+++ src/sys/util.hpp
@@ -107,3 +107,4 @@ class Signals {
 void trimString(std::string& str);
 void lowerString(std::string& str);
 std::string formatDuration(int64_t microseconds);
+std::string escapeDiscordMarkdown(const std::string& str);