+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 ## [Unreleased]
 
+## v25.11.2
+
+### Bugfixes
+
+* Fix escaping in generated Discord messages
+
 ### Features
 
+* Support multiple functions in LEAN formula generation
 * Support bitwise operations in LEAN formula generation
 * Support factorial operations in LEAN formula generation
 
 ### Enhancements
 
+* Simplify geometric progressions in formula generation
+* Simplify generated expressions for `fac` operations
+* Improve optimization of `pow`,`nrt` operations
+* Extend program stats with LEAN & PARI flags
 * Add LEAN project support for testing
 
 ## v25.10.19